[USER SECURITY] Fixed issue with JWT
Now requests can fetch the userId from the JWT and pass it to User Controllers.

+++ src/main/java/com/commitguy/backend/rest/jwt/JwtGeneratorImpl.java
@@ -40,7 +40,7 @@ public class JwtGeneratorImpl implements JwtGenerator {
                 .getBody();
 
         // Parsear datos
-        Long userId = (Long) claims.get("userId");
+        Long userId = ((Integer) claims.get("userId")).longValue();
         String nickName = (String) claims.get("nickName");
 
         return new JwtInfo(userId, nickName);
